Refactoring & Optimization (#1)
* refactoring #1

* refactoring #2

* refactoring #3

* encoding optimization

* area increased 101 x 101

* magic numbers fix #2

* logs adjustment

* directly encoding area instead of encodedarea bridge

+++ internal/terminal/screen.go
@@ -26,7 +26,7 @@ func EraseScreen(w io.Writer) {
 	fmt.Fprintf(w, "%c[2J", Esc)
 }
 
-func Erase(w io.Writer, line, column int) {
+func Erase(w io.Writer, line, column int16) {
 	MoveCursor(w, line, column)
 	fmt.Fprint(w, " ")
 }
